fix!: change EstimatedHours and ActualHours from int to float64 (#320)
## Summary

The Backlog API accepts and returns decimal values for `estimatedHours`
and `actualHours` (e.g. `2.5`). The previous `int` type caused JSON
decode errors when the API returned a fractional number, and also
prevented callers from setting fractional hour values via the option
methods.

## Changes

- `internal/model/issue.go`: Change `Issue.EstimatedHours` and
`Issue.ActualHours` from `int` to `float64`
- `internal/core/option_int.go`: Remove `WithActualHours` and
`WithEstimatedHours` (moved to float file)
- `internal/core/option_float.go`: Add `WithActualHours` and
`WithEstimatedHours` as `float64` parameters with
`positiveFloat64Option` validation
- `issue.go`: Change `Issue.EstimatedHours` and `Issue.ActualHours`
fields from `int` to `float64`; update
`IssueOptionService.WithEstimatedHours` and `WithActualHours` signatures
to `float64`
- `issue_test.go`: Update call sites from `WithActualHours(1)` /
`WithEstimatedHours(1)` to `WithActualHours(1.0)` /
`WithEstimatedHours(1.0)`

BREAKING CHANGE: `Issue.EstimatedHours` and `Issue.ActualHours` changed
from `int` to `float64`. `WithEstimatedHours` and `WithActualHours`
option methods now accept `float64` instead of `int`.

+++ issue.go
@@ -30,8 +30,8 @@ type Issue struct {
 	Milestone      []*Version
 	StartDate      Date
 	DueDate        Date
-	EstimatedHours int
-	ActualHours    int
+	EstimatedHours float64
+	ActualHours    float64
 	ParentIssueID  int
 	CreatedUser    *User
 	Created        Timestamp
@@ -434,7 +434,7 @@ type IssueOptionService struct {
 }
 
 // WithActualHours returns an option to set the `actualHours` parameter.
-func (s *IssueOptionService) WithActualHours(hours int) RequestOption {
+func (s *IssueOptionService) WithActualHours(hours float64) RequestOption {
 	return s.base.WithActualHours(hours)
 }
 
@@ -509,7 +509,7 @@ func (s *IssueOptionService) WithDueDateUntil(date string) RequestOption {
 }
 
 // WithEstimatedHours returns an option to set the `estimatedHours` parameter.
-func (s *IssueOptionService) WithEstimatedHours(hours int) RequestOption {
+func (s *IssueOptionService) WithEstimatedHours(hours float64) RequestOption {
 	return s.base.WithEstimatedHours(hours)
 }
